Update Things Page : Renamed MobileSmallBox Filename
